Replace optocoupler design module
